# src/main/resources/swagger/gestaoobrasapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gest\u00e3o Obra API",
    "description": "API para gest\u00e3o de obras e controle de usu\u00e1rios.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/obra": {
      "post": {
        "tags": [
          "Obras"
        ],
        "summary": "Cadastrar obra",
        "operationId": "criarObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObraRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Obra cadastrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida ou c\u00f3digo j\u00e1 cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Status de obra ou respons\u00e1vel n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Obras"
        ],
        "summary": "Listar obras",
        "operationId": "listarObras",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de obras",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObraResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/obra/{id}": {
      "get": {
        "tags": [
          "Obras"
        ],
        "summary": "Buscar obra por ID",
        "operationId": "buscarObraPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obra encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObraResponse"
                }
              }
            }
          },
          "404": {
            "description": "Obra n\u00e3o encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Obras"
        ],
        "summary": "Atualizar obra",
        "operationId": "atualizarObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObraRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Obra atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida ou c\u00f3digo j\u00e1 cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Obra, status ou respons\u00e1vel n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Obras"
        ],
        "summary": "Excluir obra",
        "operationId": "excluirObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Obra exclu\u00edda"
          },
          "404": {
            "description": "Obra n\u00e3o encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/obra/status": {
      "post": {
        "tags": [
          "StatusObra"
        ],
        "summary": "Cadastrar status de obra",
        "operationId": "criarStatusObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusObraRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Status de obra cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida ou c\u00f3digo j\u00e1 cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "StatusObra"
        ],
        "summary": "Listar status de obra",
        "operationId": "listarStatusObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de status de obra",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatusObraResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/obra/status/{id}": {
      "get": {
        "tags": [
          "StatusObra"
        ],
        "summary": "Buscar status de obra por ID",
        "operationId": "buscarStatusObraPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status de obra encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusObraResponse"
                }
              }
            }
          },
          "404": {
            "description": "Status de obra n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "StatusObra"
        ],
        "summary": "Atualizar status de obra",
        "operationId": "atualizarStatusObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusObraRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status de obra atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida ou c\u00f3digo j\u00e1 cadastrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Status de obra n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "StatusObra"
        ],
        "summary": "Excluir status de obra",
        "operationId": "excluirStatusObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Status de obra exclu\u00eddo"
          },
          "404": {
            "description": "Status de obra n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/obra/estoque": {
      "post": {
        "tags": [
          "EstoqueObra"
        ],
        "summary": "Cadastrar Estoque",
        "operationId": "criarEstoqueObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstoqueObraRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Estoque cadastrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstoqueObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Recurso relacionado n\u00e3o encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "EstoqueObra"
        ],
        "summary": "Listar estoques",
        "operationId": "listarEstoqueObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de estoques",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EstoqueObraResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/obra/estoque/{id}": {
      "get": {
        "tags": [
          "EstoqueObra"
        ],
        "summary": "Buscar estoque por ID",
        "operationId": "buscarEstoqueObraPorId",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Estoque encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstoqueObraResponse"
                }
              }
            }
          },
          "404": {
            "description": "Estoque n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "EstoqueObra"
        ],
        "summary": "Atualizar estoque",
        "operationId": "atualizarEstoqueObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstoqueObraRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estoque atualizado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstoqueObraResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requisi\u00e7\u00e3o inv\u00e1lida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Estoque n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EstoqueObra"
        ],
        "summary": "Excluir estoque",
        "operationId": "excluirEstoqueObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Estoque exclu\u00eddo(a)"
          },
          "404": {
            "description": "Estoque n\u00e3o encontrado(a)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/obra/movimentacao/estoque/{idObra}": {
      "get": {
        "tags": [
          "MovimentacaoEstoque"
        ],
        "summary": "Listar movimenta\u00e7\u00f5es de estoque por obra",
        "operationId": "listarMovimentacoesEstoquePorObra",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "idObra",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de movimenta\u00e7\u00f5es de estoque da obra",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MovimentacaoEstoqueResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Obra n\u00e3o encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ObraRequest": {
        "type": "object",
        "required": [
          "idStatusObra",
          "codigo",
          "nome"
        ],
        "properties": {
          "idStatusObra": {
            "type": "integer",
            "format": "int64"
          },
          "idResponsavel": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "codigo": {
            "type": "string",
            "maxLength": 50
          },
          "nome": {
            "type": "string",
            "maxLength": 150
          },
          "descricao": {
            "type": "string"
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "dataFim": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "ObraResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idStatusObra": {
            "type": "integer",
            "format": "int64"
          },
          "nomeStatusObra": {
            "type": "string"
          },
          "idResponsavel": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "chaveResponsavel": {
            "type": "string",
            "nullable": true
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "dataInicio": {
            "type": "string",
            "format": "date"
          },
          "dataFim": {
            "type": "string",
            "format": "date"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "StatusObraRequest": {
        "type": "object",
        "required": [
          "codigo",
          "nome"
        ],
        "properties": {
          "codigo": {
            "type": "string",
            "maxLength": 50
          },
          "nome": {
            "type": "string",
            "maxLength": 100
          },
          "descricao": {
            "type": "string"
          }
        }
      },
      "StatusObraResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "EstoqueObraRequest": {
        "type": "object",
        "required": [
          "idObra",
          "idMaterial",
          "quantidadeAtual",
          "quantidadeMinima"
        ],
        "properties": {
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "idMaterial": {
            "type": "integer",
            "format": "int64"
          },
          "quantidadeAtual": {
            "type": "number"
          },
          "quantidadeMinima": {
            "type": "number"
          },
          "observacao": {
            "type": "string",
            "description": "Observa\u00e7\u00e3o registrada no log de movimenta\u00e7\u00e3o"
          }
        }
      },
      "EstoqueObraResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "nomeObra": {
            "type": "string"
          },
          "idMaterial": {
            "type": "integer",
            "format": "int64"
          },
          "nomeMaterial": {
            "type": "string"
          },
          "idFornecedor": {
            "type": "integer",
            "format": "int64"
          },
          "nomeFornecedor": {
            "type": "string"
          },
          "quantidadeAtual": {
            "type": "number"
          },
          "quantidadeMinima": {
            "type": "number"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MovimentacaoEstoqueResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "idObra": {
            "type": "integer",
            "format": "int64"
          },
          "nomeObra": {
            "type": "string"
          },
          "idMaterial": {
            "type": "integer",
            "format": "int64"
          },
          "nomeMaterial": {
            "type": "string"
          },
          "idFornecedor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "nomeFornecedor": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ENTRADA",
              "SAIDA",
              "AJUSTE"
            ]
          },
          "quantidade": {
            "type": "number"
          },
          "dataMovimentacao": {
            "type": "string",
            "format": "date"
          },
          "observacao": {
            "type": "string"
          },
          "ativo": {
            "type": "boolean"
          },
          "criadoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "criadoEm": {
            "type": "string",
            "format": "date-time"
          },
          "alteradoPor": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "alteradoEm": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}